Major refactoring of models.
"talents" is now an attribute of Person.
Person has subclasses Artist, Staff and Role.
This allows artists and roles to have "talents" which means I can assign
"tenor" (for example) to both a  artist and a role, then look for artists
who are tenors to fill a role that requires a tenor.

     tenor = Singing.objects.filter(voice='tenor')
     nemorino = Role.objects.filter(name='Nemorino').first()
     avail_tenors = Artist.objects.filter(talents__contains=tenor)

Mods to admin interfaces, mainly rearranging fields and using
filter_horizontal for many-to-many selection.